Using a StdDeserializer instead of processing it in the class. Reverting changes to groupby.
