+++ core/src/main/java/org/openstack4j/model/telemetry/Statistics.java
@@ -1,7 +1,6 @@
 package org.openstack4j.model.telemetry;
 
 import java.util.Date;
-import java.util.Map;
 
 import org.openstack4j.model.ModelEntity;
 
@@ -75,5 +74,5 @@ public interface Statistics extends ModelEntity {
 	/**
    * @return The group-by of the data set
    */
-   Map<String, Object> getGroupBy();
+   String getGroupBy();
 }
