+++ core/src/main/java/org/openstack4j/openstack/telemetry/domain/CeilometerStatistics.java
@@ -1,16 +1,12 @@
 package org.openstack4j.openstack.telemetry.domain;
 
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
 import java.util.Date;
-import java.util.Map;
-import java.util.TimeZone;
 
 import org.openstack4j.model.telemetry.Statistics;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.openstack4j.openstack.common.TelemetryDateDeserializer;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.google.common.base.MoreObjects;
 
 /**
@@ -19,10 +15,8 @@ import com.google.common.base.MoreObjects;
  * @author Jeremy Unruh
  */
 public class CeilometerStatistics implements Statistics {
-   private static final Logger LOG = LoggerFactory.getLogger(CeilometerStatistics.class);
-	private static final long serialVersionUID = 1L;
 
-   private static final String MILLIS_DATE_FORMAT = "yyyy-MM-dd'T'HH:mm:ss.SSS";
+	private static final long serialVersionUID = 1L;
 
 	private Double avg;
 
@@ -30,14 +24,12 @@ public class CeilometerStatistics implements Statistics {
 
 	private Double duration;
 
+   @JsonDeserialize(using = TelemetryDateDeserializer.class)
 	@JsonProperty("duration_start")
-   private String durationStartStr;
-
-   @JsonProperty("duration_end")
-   private String durationEndStr;
-
 	private Date durationStart;
 	
+   @JsonDeserialize(using = TelemetryDateDeserializer.class)
+   @JsonProperty("duration_end")
 	private Date durationEnd;
 
 	private Double max;
@@ -46,22 +38,19 @@ public class CeilometerStatistics implements Statistics {
 
 	private Integer period;
 
+   @JsonDeserialize(using = TelemetryDateDeserializer.class)
    @JsonProperty("period_start")
-   private String periodStartStr;
+   private Date periodStart;
 
+   @JsonDeserialize(using = TelemetryDateDeserializer.class)
    @JsonProperty("period_end")
-   private String periodEndStr;
-
-	private Date periodStart;
-
 	private Date periodEnd;
 
 	private Double sum;
 
 	private String unit;
 
-   @JsonProperty("groupby")
-   private Map<String, Object> groupBy;
+   private String groupby;
 
 	/**
 	 * {@inheritDoc}
@@ -92,9 +81,6 @@ public class CeilometerStatistics implements Statistics {
 	 */
 	@Override
 	public Date getDurationStart() {
-	   if(durationStart == null){
-         durationStart = parseDate(durationStartStr);
-	   }
 		return durationStart;
 	}
 
@@ -103,9 +89,6 @@ public class CeilometerStatistics implements Statistics {
 	 */
 	@Override
 	public Date getDurationEnd() {
-      if (durationEnd == null) {
-         durationEnd = parseDate(durationEndStr);
-      }
       return durationEnd;
 	}
 
@@ -146,9 +129,6 @@ public class CeilometerStatistics implements Statistics {
 	 */
 	@Override
 	public Date getPeriodStart() {
-      if (periodStart == null) {
-         periodStart = parseDate(periodStartStr);
-      }
       return periodStart;
 	}
 
@@ -157,9 +137,6 @@ public class CeilometerStatistics implements Statistics {
 	 */
 	@Override
 	public Date getPeriodEnd() {
-      if (periodEnd == null) {
-         periodEnd = parseDate(periodEndStr);
-      }
 		return periodEnd;
 	}
 
@@ -175,39 +152,9 @@ public class CeilometerStatistics implements Statistics {
    * {@inheritDoc}
    */
   @Override
-   public Map<String, Object> getGroupBy() {
-      return groupBy;
+   public String getGroupBy() {
+      return groupby;
   }
-  
-  private Date parseDate(String date) {
-      SimpleDateFormat sdf = new SimpleDateFormat(MILLIS_DATE_FORMAT);
-      sdf.setTimeZone(TimeZone.getTimeZone("UTC"));
-    try {
-         return sdf.parse(getParseableDate(date));
-    } catch (ParseException e) {
-      LOG.warn("Error while parsing date", e);
-    }
-    return null;
-  }
-
-   /**
-    * Modifies the date string to have the expected date format ("yyyy-MM-dd'T'HH:mm:ss.SSS")
-    * 
-    * @param date
-    * @return the date with the correct format
-    */
-   private String getParseableDate(String date) {
-      // e.g. [ "2017-06-20 13:00:00" , "395000" ] or [] if no decimal
-      String[] sdate = date.split("\\.");
-      if (sdate.length > 1) {
-         return sdate[0] + "." + getExpectedDecimals(sdate[1]);
-      }
-      return date + ".000";
-   }
-
-   private String getExpectedDecimals(String decimals) {
-      return (decimals + "000").substring(0, 3);
-   }
 
    /**
     * {@inheritDoc}
@@ -216,10 +163,10 @@ public class CeilometerStatistics implements Statistics {
 	public String toString() {
 		return MoreObjects.toStringHelper(this).omitNullValues()
 				    .add("avg", avg).add("count", count).add("duration", duration)
-				    .add("durationStart", getDurationStart()).add("durationEnd", getDurationEnd())
+				    .add("durationStart", durationStart).add("durationEnd", durationEnd)
 				    .add("min", min).add("max", max).add("sum", sum).add("period", period)
-				    .add("periodStart", getPeriodStart()).add("periodEnd", getPeriodEnd()).add("unit", unit)
-				    .add("groupBy", groupBy)
+				    .add("periodStart", periodStart).add("periodEnd", periodEnd).add("unit", unit)
+				    .add("groupby", groupby)
 				    .toString();
 	}
